+++ scripts/prepare_sector_network.py
@@ -5862,7 +5862,7 @@ def lossy_bidirectional_links(n, carrier, efficiencies={}):
     carrier_i = n.links.query("carrier == @carrier").index
 
     if (
-        not any((v != 1.0) or (v >= 0) for v in efficiencies.values())
+        all((v == 1.0) for v in efficiencies.values())
         or carrier_i.empty
     ):
         return