+++ doc/release_notes.md
@@ -6,6 +6,7 @@
 <!-- Upcoming Release -->
 <!-- ================= -->
 
+* Fix: Set transmission efficiencies for `gas pipeline new` and `CO2 pipeline` links by default, to ensure consistency with other link type carriers.
 * feat: data version CSV / YAML file can be specified separately or extended by the user in the `data.version_files` config entry ([#2016](https://github.com/PyPSA/pypsa-eur/issues/2016)).
 * Fix: Resolve plotting crashes from missing `tech_colors` entries by adding `heat dsm` color and implementing upfront validation for missing keys in `plot_summary.py` ([#2108](https://github.com/PyPSA/pypsa-eur/issues/2108)).
 * feat: Make the default target rule configurable (defaults to "all" for backwards compatibility)